refactor(devices): split DeviceCard visuals into a partial file
Move the icon / glyph / accent-tile / Wave Link theming + mute-lock tooltips out of DeviceCard.cs
into DeviceCard.Visuals.cs (1188 -> 960 + 244). Only plain members are moved: the WinUI XAML
markup compiler NREs (WMC9999) when source-generated [ObservableProperty] members are spread across
partial files, so those stay in the main partial.
